+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Polyverse Boost GitHub App
 ### New Features
 - Add persistence of Installation info in AWS Dynamo
 - Add Installation deletion when app is deleted/uninstalled by User or Organization
+- Add Support for Organization installations - which have no email address
 
 ### Enhancements
 - N/A